packages/core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "コアロジックを提供するパッケージ"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "api",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "google-auth>=2.49.1",
    "requests>=2.33.0",
    "pillow>=12.1.1",
    "aiohttp>=3.13.4",
]

[dependency-groups]
dev = [
    "numpy>=2.4.4",
    "ruff>=0.15.6",
    "torch",
    "torchvision",
]

[tool.uv.sources]
api = { workspace = true }
torch = { index = "pytorch-cpu" }       # torchだけCPU版サイトを見に行く
torchvision = { index = "pytorch-cpu" } # torchだけCPU版サイトを見に行く

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
